feat: refactor broker level screening test into separate files
